Fix unit tests for patch generator
