+++ engine/src/test/java/com/google/android/fhir/sync/remote/FhirHttpDataSourceTest.kt
@@ -17,7 +17,6 @@
 import ca.uhn.fhir.context.FhirContext
 import com.google.android.fhir.ContentTypes
 import com.google.android.fhir.NetworkConfiguration
-import com.google.android.fhir.db.impl.dao.LocalChangeToken
 import com.google.android.fhir.sync.UrlUploadRequest
 import com.google.android.fhir.sync.remote.FhirHttpDataSource
 import com.google.android.fhir.sync.remote.RetrofitHttpService
@@ -86,8 +85,7 @@ internal class FhirHttpDataSourceTest {
           }
         )
       }
-    val request =
-      UrlUploadRequest(HttpVerb.POST, "Patient", patient, LocalChangeToken(listOf(1)), emptyMap())
+    val request = UrlUploadRequest(HttpVerb.POST, "Patient", patient, emptyMap())
     mockWebServer.enqueue(mockResponse)
     dataSource.upload(request)
     val recordedRequest: RecordedRequest = mockWebServer.takeRequest()
@@ -114,8 +112,7 @@ internal class FhirHttpDataSourceTest {
         HttpVerb.PUT,
         "Patient/Patient-001",
         patient,
-        LocalChangeToken(listOf(1)),
-        emptyMap()
+        emptyMap(),
       )
     mockWebServer.enqueue(mockResponse)
     dataSource.upload(request)
@@ -139,8 +136,7 @@ internal class FhirHttpDataSourceTest {
         HttpVerb.PATCH,
         "Patient/Patient-001",
         patchToApply,
-        LocalChangeToken(listOf(1)),
-        mapOf("Content-Type" to ContentTypes.APPLICATION_JSON_PATCH)
+        mapOf("Content-Type" to ContentTypes.APPLICATION_JSON_PATCH),
       )
     mockWebServer.enqueue(mockResponse)
     dataSource.upload(request)
