+++ engine/src/test/java/com/google/android/fhir/sync/upload/request/TransactionBundleGeneratorTest.kt
@@ -18,11 +18,8 @@ package com.google.android.fhir.sync.upload.request
 
 import ca.uhn.fhir.context.FhirContext
 import ca.uhn.fhir.context.FhirVersionEnum
-import com.google.android.fhir.db.impl.dao.LocalChangeToken
-import com.google.android.fhir.db.impl.entities.LocalChangeEntity
-import com.google.android.fhir.db.impl.entities.LocalChangeEntity.Type
-import com.google.android.fhir.sync.upload.patch.LocalChangeUtils
-import com.google.android.fhir.sync.upload.patch.toLocalChange
+import com.google.android.fhir.db.impl.dao.diff
+import com.google.android.fhir.sync.upload.patch.Patch
 import com.google.common.truth.Truth.assertThat
 import java.time.Instant
 import kotlinx.coroutines.runBlocking
@@ -49,85 +46,76 @@ class TransactionBundleGeneratorTest {
   fun `generateUploadRequests() should return single Transaction Bundle with 3 entries`() =
     runBlocking {
       val jsonParser = FhirContext.forCached(FhirVersionEnum.R4).newJsonParser()
-      val changes =
+      val patches =
         listOf(
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-001",
-              type = Type.INSERT,
-              payload =
-                jsonParser.encodeResourceToString(
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-001",
+            type = Patch.Type.INSERT,
+            payload =
+              jsonParser.encodeResourceToString(
+                Patient().apply {
+                  id = "Patient-001"
+                  addName(
+                    HumanName().apply {
+                      addGiven("John")
+                      family = "Doe"
+                    }
+                  )
+                }
+              ),
+            timestamp = Instant.now()
+          ),
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-002",
+            type = Patch.Type.UPDATE,
+            payload =
+              diff(
+                  jsonParser,
                   Patient().apply {
-                    id = "Patient-001"
+                    id = "Patient-002"
                     addName(
                       HumanName().apply {
-                        addGiven("John")
+                        addGiven("Jane")
+                        family = "Doe"
+                      }
+                    )
+                  },
+                  Patient().apply {
+                    id = "Patient-002"
+                    addName(
+                      HumanName().apply {
+                        addGiven("Janet")
                         family = "Doe"
                       }
                     )
                   }
-                ),
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { token = LocalChangeToken(listOf(1)) },
-          LocalChangeEntity(
-              id = 2,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-002",
-              type = Type.UPDATE,
-              payload =
-                LocalChangeUtils.diff(
-                    jsonParser,
-                    Patient().apply {
-                      id = "Patient-002"
-                      addName(
-                        HumanName().apply {
-                          addGiven("Jane")
-                          family = "Doe"
-                        }
-                      )
-                    },
-                    Patient().apply {
-                      id = "Patient-002"
-                      addName(
-                        HumanName().apply {
-                          addGiven("Janet")
-                          family = "Doe"
-                        }
-                      )
+                )
+                .toString(),
+            timestamp = Instant.now()
+          ),
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-003",
+            type = Patch.Type.DELETE,
+            payload =
+              jsonParser.encodeResourceToString(
+                Patient().apply {
+                  id = "Patient-003"
+                  addName(
+                    HumanName().apply {
+                      addGiven("John")
+                      family = "Roe"
                     }
                   )
-                  .toString(),
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { LocalChangeToken(listOf(2)) },
-          LocalChangeEntity(
-              id = 3,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-003",
-              type = Type.DELETE,
-              payload =
-                jsonParser.encodeResourceToString(
-                  Patient().apply {
-                    id = "Patient-003"
-                    addName(
-                      HumanName().apply {
-                        addGiven("John")
-                        family = "Roe"
-                      }
-                    )
-                  }
-                ),
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { LocalChangeToken(listOf(3)) }
+                }
+              ),
+            timestamp = Instant.now()
+          ),
         )
       val generator = TransactionBundleGenerator.Factory.getDefault()
-      val result = generator.generateUploadRequests(changes)
+      val result = generator.generateUploadRequests(patches)
 
       assertThat(result).hasSize(1)
       val bundleUploadRequest = result[0]
@@ -142,84 +130,75 @@ class TransactionBundleGeneratorTest {
   fun `generateUploadRequests() should return 3 Transaction Bundle with single entry each`() =
     runBlocking {
       val jsonParser = FhirContext.forCached(FhirVersionEnum.R4).newJsonParser()
-      val changes =
+      val patches =
         listOf(
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-001",
-              type = Type.INSERT,
-              payload =
-                jsonParser.encodeResourceToString(
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-001",
+            type = Patch.Type.INSERT,
+            payload =
+              jsonParser.encodeResourceToString(
+                Patient().apply {
+                  id = "Patient-001"
+                  addName(
+                    HumanName().apply {
+                      addGiven("John")
+                      family = "Doe"
+                    }
+                  )
+                }
+              ),
+            timestamp = Instant.now()
+          ),
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-002",
+            type = Patch.Type.UPDATE,
+            payload =
+              diff(
+                  jsonParser,
                   Patient().apply {
-                    id = "Patient-001"
+                    id = "Patient-002"
                     addName(
                       HumanName().apply {
-                        addGiven("John")
+                        addGiven("Jane")
+                        family = "Doe"
+                      }
+                    )
+                  },
+                  Patient().apply {
+                    id = "Patient-002"
+                    addName(
+                      HumanName().apply {
+                        addGiven("Janet")
                         family = "Doe"
                       }
                     )
                   }
-                ),
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { token = LocalChangeToken(listOf(1)) },
-          LocalChangeEntity(
-              id = 2,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-002",
-              type = Type.UPDATE,
-              payload =
-                LocalChangeUtils.diff(
-                    jsonParser,
-                    Patient().apply {
-                      id = "Patient-002"
-                      addName(
-                        HumanName().apply {
-                          addGiven("Jane")
-                          family = "Doe"
-                        }
-                      )
-                    },
-                    Patient().apply {
-                      id = "Patient-002"
-                      addName(
-                        HumanName().apply {
-                          addGiven("Janet")
-                          family = "Doe"
-                        }
-                      )
+                )
+                .toString(),
+            versionId = "v-p002-01",
+            timestamp = Instant.now()
+          ),
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-003",
+            type = Patch.Type.DELETE,
+            payload =
+              jsonParser.encodeResourceToString(
+                Patient().apply {
+                  id = "Patient-003"
+                  addName(
+                    HumanName().apply {
+                      addGiven("John")
+                      family = "Roe"
                     }
                   )
-                  .toString(),
-              versionId = "v-p002-01",
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { LocalChangeToken(listOf(2)) },
-          LocalChangeEntity(
-              id = 3,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-003",
-              type = Type.DELETE,
-              payload =
-                jsonParser.encodeResourceToString(
-                  Patient().apply {
-                    id = "Patient-003"
-                    addName(
-                      HumanName().apply {
-                        addGiven("John")
-                        family = "Roe"
-                      }
-                    )
-                  }
-                ),
-              versionId = "v-p003-01",
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
-            .apply { LocalChangeToken(listOf(3)) }
+                }
+              ),
+            versionId = "v-p003-01",
+            timestamp = Instant.now()
+          ),
         )
       val generator =
         TransactionBundleGenerator.Factory.getGenerator(
@@ -228,7 +207,7 @@ class TransactionBundleGeneratorTest {
           1,
           true
         )
-      val result = generator.generateUploadRequests(changes)
+      val result = generator.generateUploadRequests(patches)
 
       // Exactly 3 Requests are generated
       assertThat(result).hasSize(3)
@@ -247,21 +226,19 @@ class TransactionBundleGeneratorTest {
   @Test
   fun `generate() should return Bundle Entry without if-match when useETagForUpload is false`() =
     runBlocking {
-      val changes =
+      val patches =
         listOf(
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-002",
-              type = Type.UPDATE,
-              payload = "[]",
-              versionId = "patient-002-version-1",
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-002",
+            type = Patch.Type.UPDATE,
+            payload = "[]",
+            versionId = "patient-002-version-1",
+            timestamp = Instant.now()
+          )
         )
       val generator = TransactionBundleGenerator.Factory.getDefault(useETagForUpload = false)
-      val result = generator.generateUploadRequests(changes)
+      val result = generator.generateUploadRequests(patches)
 
       assertThat(result.first().resource.entry.first().request.ifMatch).isNull()
     }
@@ -269,21 +246,19 @@ class TransactionBundleGeneratorTest {
   @Test
   fun `generate() should return Bundle Entry with if-match when useETagForUpload is true`() =
     runBlocking {
-      val changes =
+      val patches =
         listOf(
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-002",
-              type = Type.UPDATE,
-              payload = "[]",
-              versionId = "patient-002-version-1",
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-002",
+            type = Patch.Type.UPDATE,
+            payload = "[]",
+            versionId = "patient-002-version-1",
+            timestamp = Instant.now()
+          )
         )
       val generator = TransactionBundleGenerator.Factory.getDefault(useETagForUpload = true)
-      val result = generator.generateUploadRequests(changes)
+      val result = generator.generateUploadRequests(patches)
 
       assertThat(result.first().resource.entry.first().request.ifMatch)
         .isEqualTo("W/\"patient-002-version-1\"")
@@ -292,31 +267,27 @@ class TransactionBundleGeneratorTest {
   @Test
   fun `generate() should return Bundle Entry without if-match when the LocalChangeEntity has no versionId`() =
     runBlocking {
-      val changes =
+      val patches =
         listOf(
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-002",
-              type = Type.UPDATE,
-              payload = "[]",
-              versionId = "",
-              timestamp = Instant.now()
-            )
-            .toLocalChange(),
-          LocalChangeEntity(
-              id = 1,
-              resourceType = ResourceType.Patient.name,
-              resourceId = "Patient-003",
-              type = Type.UPDATE,
-              payload = "[]",
-              versionId = null,
-              timestamp = Instant.now()
-            )
-            .toLocalChange()
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-002",
+            type = Patch.Type.UPDATE,
+            payload = "[]",
+            versionId = "",
+            timestamp = Instant.now()
+          ),
+          Patch(
+            resourceType = ResourceType.Patient.name,
+            resourceId = "Patient-003",
+            type = Patch.Type.UPDATE,
+            payload = "[]",
+            versionId = null,
+            timestamp = Instant.now()
+          ),
         )
       val generator = TransactionBundleGenerator.Factory.getDefault(useETagForUpload = true)
-      val result = generator.generateUploadRequests(changes)
+      val result = generator.generateUploadRequests(patches)
 
       assertThat(result.first().resource.entry[0].request.ifMatch).isNull()
       assertThat(result.first().resource.entry[1].request.ifMatch).isNull()
